reverted the config class
it wasn't really needed
